Fix small bug in NextFit
Fix wrap around functionality. The issue was
that after freeing a Block and allocating space
for a new Block, the nextFit search start point was
not updated correctly.

It was due to the Block used attribute not being
set to true after finding one Block that was free
to use.